Dependency cleanup

+++ mitosheet/mitosheet/tests/steps_performers/test_filter.py
@@ -52,7 +52,7 @@ FILTER_TESTS = [
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
         FC_NUMBER_EXACTLY,
         10,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ),
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -64,7 +64,7 @@ FILTER_TESTS = [
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
         FC_NUMBER_GREATER,
         10,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -82,7 +82,7 @@ FILTER_TESTS = [
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
         FC_NUMBER_GREATER_THAN_OR_EQUAL,
         10,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -100,7 +100,7 @@ FILTER_TESTS = [
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
         FC_NUMBER_LESS,
         0,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
@@ -118,7 +118,7 @@ FILTER_TESTS = [
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
         FC_NUMBER_LESS_THAN_OR_EQUAL,
         0,
-        pd.DataFrame(columns=['A']),
+        pd.DataFrame(columns=['A'], dtype='int'),
     ), 
     (
         pd.DataFrame(data={'A': [1, 2, 3, 4, 5, 6]}),
